requirements.txt
flask
pandas
matplotlib
ta
python-binance